Refactoring: MachineConfigurationBuilder is now an interface and the MachineJsonImporter no longer works with adding loaded data to lists directly
