+++ src/main/java/de/uni_hannover/sra/minimax_simulator/io/importer/json/MachineJsonImporter.java
@@ -4,6 +4,7 @@ import de.uni_hannover.sra.minimax_simulator.Main;
 import de.uni_hannover.sra.minimax_simulator.io.importer.ProjectImportException;
 import de.uni_hannover.sra.minimax_simulator.model.configuration.MachineConfiguration;
 import de.uni_hannover.sra.minimax_simulator.model.configuration.MachineConfigurationBuilder;
+import de.uni_hannover.sra.minimax_simulator.model.configuration.MinimaxConfigurationBuilder;
 import de.uni_hannover.sra.minimax_simulator.model.configuration.alu.AluOperation;
 import de.uni_hannover.sra.minimax_simulator.model.configuration.mux.*;
 import de.uni_hannover.sra.minimax_simulator.model.configuration.register.RegisterExtension;
@@ -48,15 +49,15 @@ class MachineJsonImporter extends Importer {
 		ResourceBundleLoader resourceLoader = Main.getResourceLoader();
 		TextResource registerTextResource = resourceLoader.getTextResource("register");
 
-		MachineConfigurationBuilder mb = new MachineConfigurationBuilder();
+		MachineConfigurationBuilder mb = new MinimaxConfigurationBuilder();
 		mb.addDefaultBaseRegisters(registerTextResource);
-		mb.getAluOperations().addAll(loadAlu(alu));
-		mb.getRegisterExtensions().addAll(loadRegisters(registers));
+		mb.addAluOperations(loadAlu(alu));
+		mb.addRegisterExtensions(loadRegisters(registers));
 
 		JSONArray muxInputList = machine.getJSONArray("muxInputs");
 		for (int i = 0; i < muxInputList.length(); i++) {
 			MuxType type = get(MuxType.class, muxInputList.getJSONObject(i).getString("muxType"));
-			mb.getSelectedMuxInputs(type).addAll(loadMuxInputs(muxInputList.getJSONObject(i).getJSONArray("input"), mb.getAllowedMuxInputs()));
+			mb.addMuxInputs(type, loadMuxInputs(muxInputList.getJSONObject(i).getJSONArray("input"), mb.getAllowedMuxInputs()));
 		}
 
 		return mb.build();
