+++ src/main/java/de/uni_hannover/sra/minimax_simulator/model/signal/SignalConfiguration.java
@@ -15,7 +15,6 @@ public interface SignalConfiguration {
 	 * @return
 	 *          a list of the {@code SignalType}s
 	 */
-	// TODO: use immutable lists for other interfaces, too
 	public ImmutableList<SignalType> getSignalTypes();
 
 	/**
