+++ src/test/java/de/uni_hannover/sra/minimax_simulator/ui/editor/AluTableModelTest.java
@@ -1,13 +1,6 @@
 package de.uni_hannover.sra.minimax_simulator.ui.editor;
 
-import de.uni_hannover.sra.minimax_simulator.Main;
-import de.uni_hannover.sra.minimax_simulator.model.configuration.alu.AluOperation;
-
 import org.junit.Test;
-
-import de.uni_hannover.sra.minimax_simulator.model.configuration.MachineConfiguration;
-import de.uni_hannover.sra.minimax_simulator.model.configuration.MachineConfigurationBuilder;
-import de.uni_hannover.sra.minimax_simulator.resources.ResourceBundleLoader;
 //import de.uni_hannover.sra.minimax_simulator.ui.tabs.machine.overview.model.AluTableModel;
 
 public class AluTableModelTest
